refactor: extract RootContext to separate file
Split RootContext from actor/system.rs into actor/root_context.rs
for better code organization. Each public type now has its own
module file following Rust 2018 conventions.
