+++ modules/actor-core/src/actor.rs
@@ -4,13 +4,15 @@ mod context;
 mod message_envelope;
 mod props;
 mod system;
+mod root_context;
 
 pub use actor_ref::ActorRef;
 pub use behavior::{ActorAdapter, Behavior};
 pub use context::Context;
 pub use message_envelope::MessageEnvelope;
 pub use props::Props;
-pub use system::{ActorSystem, RootContext};
+pub use system::{ActorSystem};
+pub use root_context::RootContext;
 
 #[cfg(test)]
 mod tests;
