+++ modules/actor-core/src/actor/system.rs
@@ -4,7 +4,7 @@ use crate::guardian::AlwaysRestart;
 use crate::system::{InternalActorSystem, InternalRootContext};
 use crate::{MailboxRuntime, PriorityEnvelope};
 use nexus_utils_core_rs::{Element, QueueError};
-
+use crate::actor::root_context::RootContext;
 use super::{ActorRef, MessageEnvelope, Props};
 
 pub struct ActorSystem<U, R, Strat = AlwaysRestart>
@@ -79,46 +79,5 @@ where
 
   pub async fn dispatch_next(&mut self) -> Result<(), QueueError<PriorityEnvelope<MessageEnvelope<U>>>> {
     self.inner.dispatch_next().await
-  }
-}
-
-pub struct RootContext<'a, U, R, Strat>
-where
-  U: Element,
-  R: MailboxRuntime + Clone + 'static,
-  R::Queue<PriorityEnvelope<MessageEnvelope<U>>>: Clone,
-  R::Signal: Clone,
-  Strat: crate::guardian::GuardianStrategy<MessageEnvelope<U>, R>, {
-  inner: InternalRootContext<'a, MessageEnvelope<U>, R, Strat>,
-}
-
-impl<'a, U, R, Strat> RootContext<'a, U, R, Strat>
-where
-  U: Element,
-  R: MailboxRuntime + Clone,
-  R::Queue<PriorityEnvelope<MessageEnvelope<U>>>: Clone,
-  R::Signal: Clone,
-  Strat: crate::guardian::GuardianStrategy<MessageEnvelope<U>, R>,
-{
-  pub fn spawn(
-    &mut self,
-    props: Props<U, R>,
-  ) -> Result<ActorRef<U, R>, QueueError<PriorityEnvelope<MessageEnvelope<U>>>> {
-    let actor_ref = self.inner.spawn(props.into_inner())?;
-    Ok(ActorRef::new(actor_ref))
-  }
-
-  #[deprecated(since = "3.1.0", note = "dispatch_next / run_until を使用してください")]
-  pub fn dispatch_all(&mut self) -> Result<(), QueueError<PriorityEnvelope<MessageEnvelope<U>>>> {
-    #[allow(deprecated)]
-    self.inner.dispatch_all()
-  }
-
-  pub async fn dispatch_next(&mut self) -> Result<(), QueueError<PriorityEnvelope<MessageEnvelope<U>>>> {
-    self.inner.dispatch_next().await
-  }
-
-  pub fn raw(&mut self) -> &mut InternalRootContext<'a, MessageEnvelope<U>, R, Strat> {
-    &mut self.inner
   }
 }
